refactor: extract large css to style modules
